feat: add information of  annotation setting  when retrieve app

+++ internal/api-server/domain/chat/entity/biz_entity/entity_annotation_setting.go
@@ -17,7 +17,7 @@ type CollectionBingDetail struct {
 type AnnotationSettingWithBinding struct {
 	ID                      string                `json:"id"`
 	AppID                   string                `json:"app_id"`
-	ScoreThreshold          float64               `json:"score_threshold"`
+	ScoreThreshold          float32               `json:"score_threshold"`
 	CollectionBindingID     string                `json:"collection_binding_id"`
 	CreatedUserID           string                `json:"created_user_id"`
 	CreatedAt               int64                 `json:"created_at"`
